add filter mock logic template
