+++ src/main.js
@@ -6,7 +6,7 @@ import {createCardsSortTemplate as sort} from './view/tasks-sort';
 // import {createTaskCardTemplate as task} from './view/task';
 import {createLoadButtonTemplate as loadMoreBtn} from './view/load-more-btn';
 import {generateTask, createTaskTemplate} from "./mock/task";
-import {createTaskFormTemplate} from "./mock/edit-task";
+import {createTaskFormTemplate} from "./mock/task-edit";
 
 const TASK_AMOUNT = 10;
 const content = document.querySelector(`.main`);
@@ -15,6 +15,7 @@ const boardContainer = content.querySelector(`.board`);
 const taskListContainer = boardContainer.querySelector(`.board__tasks`);
 
 const tasks = new Array(TASK_AMOUNT).fill().map(generateTask);
+console.log(tasks);
 
 const addTasks = (container, position) => {
   const length = tasks.length - 1;
